requirements.txt
# Exchange API
ccxt>=4.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Configuration
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Code quality
black>=23.0.0
ruff>=0.0.280
mypy>=1.5.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
